feat(Machine): break responsibilties of Machine class down into Micro and Macro machine classes
